refactor: main 소스 코드를 MVC 및 도메인 기반 패키지로 분리
