+++ src/main/java/lotto/Application.java
@@ -1,5 +1,10 @@
 package lotto;
 
+import lotto.controller.LottoGameController;
+import lotto.service.LottoService;
+import lotto.view.InputView;
+import lotto.view.OutputView;
+
 public class Application {
 
     public static void main(String[] args) {
