+++ src/main/java/lotto/service/LottoService.java
@@ -1,6 +1,10 @@
-package lotto;
+package lotto.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import lotto.domain.Lotto;
+import lotto.domain.LottoResult;
+import lotto.domain.Rank;
+
 import java.util.ArrayList;
 import java.util.List;
