+++ src/main/java/lotto/util/Validator.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.util;
+
+import lotto.domain.Lotto;
 
 import java.util.ArrayList;
 import java.util.List;
