+++ src/test/java/lotto/RankTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
